requirements.txt
# Core LangChain stack
langchain>=0.2
langchain-openai>=0.1
openai>=1.0

# API server
fastapi>=0.110
uvicorn>=0.27

# UI
gradio>=4.0

# HTTP client (used by ui_gradio_api.py)
requests>=2.31

# Environment variables
python-dotenv>=1.0

# PostgreSQL driver
psycopg2-binary>=2.9

# Embedding model
sentence-transformers>=2.7

# pgvector Python helper (optional – used for type-safe vector literals)
# pgvector>=0.2
